requirements.txt
affine==2.2.2
alabaster==0.7.12
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
bmi-python==0.2.7
certifi==2019.3.9
cffi==1.12.3
cftime==1.0.3.4
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cloudpickle==1.1.1
cmocean==2.0
colorama==0.4.1
cryptography==2.7
cycler==0.10.0
cytoolz==0.9.0.1
dask==1.2.2
decorator==4.4.0
defusedxml==0.6.0
descartes==1.1.0
docopt==0.6.1
docopts==0.6.1
docutils==0.14
entrypoints==0.3
GDAL==2.3.3
geojson==2.4.1
idna==2.8
imageio==2.5.0
imagesize==1.1.0
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.20
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
logutils==0.3.5
Mako==1.0.10
MarkupSafe==1.1.1
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
mmi==0.7.0
more-itertools==7.0.0
nbconvert==5.5.0
nbformat==4.4.0
netCDF4==1.4.2
networkx==2.3
numpy==1.14.6
numpydoc==0.9.1
olefile==0.46
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pathlib==1.0.1
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
psutil==5.6.2
pycodestyle==2.5.0
pycparser==2.19
pycpt==0.2.0
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.14.11
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
pywin32==223
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.5.1
QtPy==1.7.1
rainbow-logging-handler==2.2.2
rasterio==1.0.21
requests==2.22.0
rope==0.14.0
-e git+https://github.com/openearth/sandbox-fm.git@f9bbe06153bba7aa8bb598f1341e7e94e3af73f8#egg=sandbox_fm
scikit-image==0.15.0
scipy==1.2.1
Shapely==1.6.4.post1
six==1.12.0
snowballstemmer==1.2.1
snuggs==1.4.6
Sphinx==2.1.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==3.3.4
spyder-kernels==0.4.4
testpath==0.4.2
toolz==0.9.0
tornado==6.0.2
tqdm==4.32.2
traitlets==4.3.2
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.1
